introduce base.py instead of core.py
