+++ cou/apps/subordinate.py
@@ -15,7 +15,7 @@
 import logging
 from typing import Optional
 
-from cou.apps.core import OpenStackApplication
+from cou.apps.base import OpenStackApplication
 from cou.apps.factory import AppFactory
 from cou.steps import UpgradeStep
 from cou.utils.openstack import SUBORDINATES, OpenStackRelease
